+++ example/views/applications.py
@@ -4,12 +4,14 @@ from example.views import BaseView
 class ApplicationsView(BaseView):
 
     def index(self):
-        return [
-            dict(id=1, name='App 1'),
-            dict(id=2, name='App 2'),
-            dict(id=3, name='App 3'),
-            dict(id=4, name='App 4'),
-            ]
+        return {
+            'data': [
+                dict(id=1, name='App 1'),
+                dict(id=2, name='App 2'),
+                dict(id=3, name='App 3'),
+                dict(id=4, name='App 4'),
+                ],
+            }
 
     def update(self, id):
         return {}